chore(utils): reflecting changes in ee utils
reflect try_decode_base64 moved to utils (since
https://github.com/Kong/kong-ee/pull/3872)
